Remove all the v1beta1 stuff
